requirements.txt

numpy==1.24.4
opencv-python==4.8.1.78
pillow==10.0.1
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
pycocotools==2.0.7
tqdm==4.66.1
flask==3.0.0


